enemies move to houses
enemy moves from spawn point to correlating house

updated size of the fort to fix this, as the enemy kinda moves in an
arc, and it ran into a fort wall. moved the wall further away from
building to prevent it

I removed all starting enemies but 3, so no enemies show up until
spawned

starting enemies no longer move.

to give an enemy a target, sendmessage("getNextTarget", string) where
string is the tag of the target