Damage is better. Needs to be called by gameMaster.
I just made the game master call it every frame on all the units on a
given team.
Houses have health now, but don't take damage for some reason. I gave
them a lot of health so that way it takes a while for the enemies to
blow them up.